wip: refactoring and reorganizaton of the code and files
