+++ 5.35.0/themeTypographyMigration/createThemeUpgradeBackup.ts
@@ -15,10 +15,7 @@ export type ThemeBackupResult = {
 export const createThemeUpgradeBackup = async (context: Context): Promise<ThemeBackupResult> => {
     const currentProjectVersion = findVersion(path.join(context.project.root, "package.json"));
 
-    log.info(
-        "Backing up %s app theme folder and generating latest code...",
-        "apps/theme"
-    );
+    log.info("Backing up %s app theme folder and generating latest code...", "apps/theme");
 
     const results = [false];
 
@@ -49,9 +46,7 @@ export const createThemeUpgradeBackup = async (context: Context): Promise<ThemeB
     }
 
     if (!results.includes(false)) {
-        log.success(
-            "%s app theme folder successfully backed up and moved to app/_theme_backup"
-        );
+        log.success("%s app theme folder successfully backed up and moved to app/_theme_backup");
         return {
             isSuccessful: true
         };
